Рециклер работает.

+++ app/src/main/java/pro/butovanton/noface/Models/Massage.kt
@@ -1,8 +1,9 @@
 package pro.butovanton.noface.Models
 
+import android.text.Editable
 import com.google.firebase.database.IgnoreExtraProperties
 
 @IgnoreExtraProperties
 data class Massage(
-    var time: Int,
+    var time: Long,
     var text: String )
